Storage pluggable refactor
