+++ packages/storage/src/Providers/index.ts
@@ -1,0 +1,3 @@
+import AWSS3Provider from './AWSS3Provider';
+
+export { AWSS3Provider };
