+++ packages/storage/src/index.ts
@@ -12,6 +12,7 @@
  */
 
 import StorageClass from './Storage';
+import { StorageProvider } from './types';
 
 import Amplify, { ConsoleLogger as Logger } from '@aws-amplify/core';
 
@@ -21,17 +22,17 @@ let _instance: StorageClass = null;
 
 if (!_instance) {
     logger.debug('Create Storage Instance');
-    _instance = new StorageClass(null);
-    _instance.vault = new StorageClass({ level: 'private' });
+    _instance = new StorageClass();
+    // _instance.vault = new StorageClass({ level: 'private' });
 
-    const old_configure = _instance.configure;
-    _instance.configure = (options) => {
-        logger.debug('configure called');
-        old_configure.call(_instance, options);
+    // const old_configure = _instance.configure;
+    // _instance.configure = (options) => {
+    //     logger.debug('configure called');
+    //     old_configure.call(_instance, options);
 
-        const vault_options = Object.assign({}, options, { level: 'private' });
-        _instance.vault.configure(vault_options);
-    };
+    //     const vault_options = Object.assign({}, options, { level: 'private' });
+    //     _instance.vault.configure(vault_options);
+    // };
 }
 
 const Storage = _instance;
@@ -39,3 +40,5 @@ Amplify.register(Storage);
 
 export default Storage;
 export { StorageClass };
+export { StorageProvider };
+export * from './Providers';
