+++ packages/storage/src/types/Storage.ts
@@ -10,13 +10,13 @@
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-
+import { AWS } from '@aws-amplify/core';
 /**
  * Storage instance options
  */
 export interface StorageOptions {
     bucket?: string;
     region?: string;
-    credentials?: object;
     level?: string;
+    credentials?: AWS.Credentials & AWS.CognitoIdentityCredentials;
 }
