+++ packages/storage/src/types/index.ts
@@ -12,3 +12,4 @@
  */
 
 export * from './Storage';
+export * from './Provider';
